chore(v0.7.5): cleanup + guardrails bundle (8 audit items)
Implements the 8 quick-wins from the 2026-05-22 engineering plan
(synthesizes 4 independent codebase reviews). All low-risk individually;
bundled into one PR to limit review surface across small mechanical
fixes. v0.7.5 release mechanics + the larger failover-engine rebuild
ship in a separate follow-up PR.

## Changes

### Test hygiene
- Delete `packages/fox-overlay/tests/test_providers_patch.py` — 201-line
  test file imported `fox_overlay.webui_patches.providers`, a module
  retired in v0.6.2 (#269). Either CI silently skipped it (a real gap)
  or it failed and nobody noticed.

### Anchor-drift fails loud (closes a real silent-shipping risk)
- `bootstrap.py`: stop swallowing `AssertionError` from `webui_patches.
  apply_all()`. Anchor-or-signature drift is the entire point of
  `_helpers.substitute_function` failing loudly; demoting it to a
  WARNING shipped silently-broken Fox to users. Now: `AssertionError`
  re-raises (aborts boot, CI catches), other exceptions use
  `_log.exception` for full traceback (still degrades that one patch).

### Offline-resilient image management
- `scripts/install.sh`: drop the pre-pull `docker rmi`. The old
  rationale ("force a fresh pull from registry") was wrong — `docker
  pull` for a tagged ref already checks the manifest and pulls a new
  digest if `:stable` was re-tagged. The rmi discarded the offline
  fallback for no benefit.
- `electron/src/docker-manager.js`: same fix. Plus updated comment
  on the failure mode: pull failure with local image present now
  degrades gracefully (no updates this launch) instead of stranding
  the user.

### Rollback escape hatch
- `scripts/install.sh` + `electron/src/docker-manager.js`: respect
  `FITB_IMAGE` env override. Users stranded by a bad release can
  roll back via `FITB_IMAGE=ghcr.io/.../cloud:v0.7.3 ./install.sh`
  or by setting the env var before launching the desktop app.
  Default unchanged.

### Log rotation
- `integration/supervisord.conf`: add
  `stdout_logfile_maxbytes=10MB, stdout_logfile_backups=3` (and
  stderr equivalents) to all 6 `[program:*]` blocks. Caps per-service
  log growth at ~240MB worst case across the container (was unbounded
  in practice — supervisord's per-program defaults are 50MB × 10).

### Duplication removal
- New `fox_overlay/_substitute.py` holds the canonical
  `substitute_function` + `substitute_method`. Both `_helpers.py`
  files (webui side + agent side) are now 3-line re-export shims.
  The two copies had already drifted once (webui added
  `substitute_method`; agent didn't); this collapse prevents the
  next drift.

### Docstring trim
- `webui_patches/streaming.py`: 60-line "Phase 8 refresh history"
  → focused docstring on what the patch does today (+ one-line note
  on the dropped hunks). Git history has the full context for anyone
  who wants it.
- `webui_patches/__init__.py`: shorten the v0.6.2 / Phase-8 retirement
  essays to single-line notes.

### New CI gate: Option B diff guard
- `.github/workflows/option-b-diff-guard.yml`: fail any PR titled
  `bump(upstream): …` whose diff touches anything outside
  `forks/hermes-{agent,webui}` (submodule pointers) +
  `packages/fox-overlay/versions.toml`. Closes the last hole in the
  Option B auto-bump trust model — pre-v0.7.5, a typo-prefixed Fox-code
  PR could silently ship arbitrary code to `:stable`.

## Verification

- 125 / 125 tests pass (1 skipped) — same as before deletion of dead
  test file; refactored `_helpers.py` imports verified by existing
  `test_config_patch.py` + `test_streaming_patch.py` + `test_bootstrap.py`
- `check-overlay-basis.sh` clean against current pin (webui v0.51.107)
- Net diff: -459 lines across 9 files (large cleanup, small additions)
- New `option-b-diff-guard.yml` validates as parseable YAML

## Out of scope

- Failover engine rebuild for `#303` symptom 3 — separate PR (v0.7.5 PR B)
- VERSION / CHANGELOG / smoke-checklist updates — land with the
  release-mechanics PR alongside PR B
