+++ packages/integration/supervisord.conf
@@ -29,6 +29,14 @@ autostart=true
 autorestart=true
 stdout_logfile=/data/logs/tailscaled.log
 stderr_logfile=/data/logs/tailscaled.err
+; v0.7.5: cap per-service logs at 10MB × 4 (matches [supervisord]'s own
+; logfile cap). Supervisord's per-program defaults are 50MB × 10 — too
+; generous for a self-hosted container where /data is on the user's host
+; disk and a chatty service could silently fill it.
+stdout_logfile_maxbytes=10MB
+stdout_logfile_backups=3
+stderr_logfile_maxbytes=10MB
+stderr_logfile_backups=3
 priority=10
 
 ; ── tailscale operator watchdog (FITB#127) ────────────────────────────────────
@@ -47,6 +55,10 @@ autostart=true
 autorestart=true
 stdout_logfile=/data/logs/ts-operator-watchdog.log
 stderr_logfile=/data/logs/ts-operator-watchdog.err
+stdout_logfile_maxbytes=10MB
+stdout_logfile_backups=3
+stderr_logfile_maxbytes=10MB
+stderr_logfile_backups=3
 priority=15
 
 ; ── qdrant ───────────────────────────────────────────────────────────────────
@@ -57,6 +69,10 @@ autostart=true
 autorestart=true
 stdout_logfile=/data/logs/qdrant.log
 stderr_logfile=/data/logs/qdrant.err
+stdout_logfile_maxbytes=10MB
+stdout_logfile_backups=3
+stderr_logfile_maxbytes=10MB
+stderr_logfile_backups=3
 priority=20
 
 ; ── hermes gateway ────────────────────────────────────────────────────────────
@@ -69,6 +85,10 @@ autostart=true
 autorestart=true
 stdout_logfile=/data/logs/hermes-gateway.log
 stderr_logfile=/data/logs/hermes-gateway.err
+stdout_logfile_maxbytes=10MB
+stdout_logfile_backups=3
+stderr_logfile_maxbytes=10MB
+stderr_logfile_backups=3
 environment=HOME="/app",PYTHONPATH="/data/apps/hermes-agent",PATH="/usr/local/bin:/usr/bin:/bin",HERMES_HOME="/data/data/hermes",BRAVE_API_KEY="__BRAVE_API_KEY__"
 priority=30
 
@@ -80,6 +100,10 @@ autostart=true
 autorestart=true
 stdout_logfile=/data/logs/hermes-webui.log
 stderr_logfile=/data/logs/hermes-webui.err
+stdout_logfile_maxbytes=10MB
+stdout_logfile_backups=3
+stderr_logfile_maxbytes=10MB
+stderr_logfile_backups=3
 ; QA fix v0.4.7-WaveI: HERMES_WEBUI_STATE_DIR was unset, so config.py
 ; defaulted to HOME/.hermes/webui = /app/.hermes/webui. /app is INSIDE
 ; the image and is wiped on every container restart — meaning every
@@ -136,4 +160,8 @@ autorestart=true
 startretries=2
 stdout_logfile=/data/logs/llama-server.log
 stderr_logfile=/data/logs/llama-server.err
+stdout_logfile_maxbytes=10MB
+stdout_logfile_backups=3
+stderr_logfile_maxbytes=10MB
+stderr_logfile_backups=3
 priority=50
